+++ database/connection.py
@@ -1,0 +1,39 @@
+import os
+
+from config import settings
+from sqlalchemy import create_engine
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm import sessionmaker
+
+# Configuración de la base de datos
+DATABASE_URL = os.getenv("DATABASE_URL")
+
+engine = create_engine(
+    settings.DATABASE_URL,
+    pool_pre_ping=True,
+    pool_recycle=300,
+    echo=False,
+)
+
+SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
+
+Base = declarative_base()
+
+
+def get_db():
+    """Dependency to get database session."""
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
+
+
+def create_tables():
+    """Create all tables."""
+    Base.metadata.create_all(bind=engine)
+
+
+def drop_tables():
+    """Drop all tables (use with caution)."""
+    Base.metadata.drop_all(bind=engine)
