+++ src/entities/__init__.py
@@ -1,0 +1,8 @@
+from .cita import Cita
+from .diagnostico import Diagnostico
+from .enfermera import Enfermera
+from .factura import Factura
+from .medico import Medico
+from .paciente import Paciente
+
+__all__ = ["Paciente", "Medico", "Enfermera", "Cita", "Diagnostico", "Factura"]
