+++ src/routers/Diagnosticos.py
